requirements.txt
aiohttp==3.9.1
discord.py==2.0.1
python-dotenv==1.0.1
pytz==2023.3
Requests==2.31.0
